move vendor_scripts_view specs
