+++ lib/assets/test/spec/cartodb/common/vendor_scripts_view.spec.js
@@ -1,3 +1,4 @@
+var _ = require('underscore');
 var cdb = require('cartodb.js-v3');
 var VendorScriptsView = require('../../../../javascripts/cartodb/common/vendor_scripts_view');
 
@@ -13,14 +14,17 @@ var CONFIG = {
   intercom_app_id: 'intercom_app_id'
 };
 
+var USER = {
+  username: 'pepe',
+  base_url: 'http://pepe.carto.com',
+  email: 'e00000002@d00000002.com',
+  account_type: 'FREE',
+  intercom: true
+};
+
 describe('common/vendor_scripts_view', function () {
   beforeEach(function () {
-    this.user = new cdb.admin.User({
-      username: 'pepe',
-      base_url: 'http://pepe.carto.com',
-      email: 'e00000002@d00000002.com',
-      account_type: 'FREE'
-    });
+    this.user = new cdb.admin.User(USER);
     spyOn(this.user, 'featureEnabled').and.returnValue(true);
 
     this.view = new VendorScriptsView({
@@ -36,20 +40,20 @@ describe('common/vendor_scripts_view', function () {
     this.view.render();
 
     expect(this.view.template).toHaveBeenCalledWith({
+      assetsVersion: '1.0.0',
+      googleAnalyticsDomain: 'carto.com',
+      googleAnalyticsMemberType: 'FREE',
+      googleAnalyticsUa: 'UA-20934186-25',
+      hubspotEnabled: true,
+      hubspotIds: '{}',
+      hubspotToken: 'yourtoken',
+      intercomAppId: 'intercom_app_id',
+      intercomEnabled: true,
+      trackjsAppKey: 'trackjs_app_key',
       trackjsCustomer: 'trackjs_customer',
       trackjsEnabled: true,
-      trackjsAppKey: 'trackjs_app_key',
-      assetsVersion: '1.0.0',
-      userName: 'pepe',
-      ua: 'UA-20934186-25',
-      domain: 'carto.com',
-      memberType: 'FREE',
-      hubspotEnabled: true,
-      hubspotToken: 'yourtoken',
-      hubspotIds: '{}',
-      intercomEnabled: true,
-      intercomAppId: 'intercom_app_id',
-      userEmail: 'e00000002@d00000002.com'
+      userEmail: 'e00000002@d00000002.com',
+      userName: 'pepe'
     });
   });
 
@@ -60,6 +64,47 @@ describe('common/vendor_scripts_view', function () {
     this.view.render();
 
     expect(this.view.$el.html).toHaveBeenCalledWith('<div>foo bar!</div>');
+  });
+
+  it('should not render hubspot script if it is not enabled', function () {
+    var CONFIG_NO_VENDOR = _.extend(CONFIG, {
+      hubspot_enabled: false,
+      trackjs_enabled: false,
+      google_analytics_ua: null
+    });
+
+    var USER_NO_VENDOR = _.extend(USER, {
+      intercom: false
+    });
+
+    this.user = new cdb.admin.User(USER_NO_VENDOR);
+
+    this.view = new VendorScriptsView({
+      config: CONFIG_NO_VENDOR,
+      assetsVersion: '1.0.0',
+      user: this.user
+    });
+
+    spyOn(this.view, 'template');
+
+    this.view.render();
+
+    expect(this.view.template).toHaveBeenCalledWith({
+      assetsVersion: '1.0.0',
+      googleAnalyticsDomain: 'carto.com',
+      googleAnalyticsMemberType: 'FREE',
+      googleAnalyticsUa: null,
+      hubspotEnabled: false,
+      hubspotIds: '{}',
+      hubspotToken: 'yourtoken',
+      intercomAppId: 'intercom_app_id',
+      intercomEnabled: false,
+      trackjsAppKey: 'trackjs_app_key',
+      trackjsCustomer: 'trackjs_customer',
+      trackjsEnabled: false,
+      userEmail: 'e00000002@d00000002.com',
+      userName: 'pepe'
+    });
   });
 
   it('should have no leaks', function () {
